Refactors the support for dynamic IdPs
Initially the functionallity for retrieving dynamically the IdP
entryPoint was implemented in the passport-saml library. This commit
rolls back the changes passport-saml and re-implemments the dynamic IdP
functionallity in the exress app.

This implementation is more flexible in changing passport-saml
configuration parameters in a dynamic manner and removes the external
dependency to the IdP model in the the passport-saml libradry
